Revert last 2 changes.
This reverts commit 5dd334270e0d4dc20f49cab79ee07f8b1170bf6a.
