+++ hedera-mirror-grpc/src/main/java/com/hedera/mirror/grpc/controller/ConsensusController.java
@@ -16,15 +16,21 @@
 
 package com.hedera.mirror.grpc.controller;
 
+import com.google.protobuf.InvalidProtocolBufferException;
 import com.hedera.mirror.api.proto.ConsensusTopicQuery;
 import com.hedera.mirror.api.proto.ConsensusTopicResponse;
 import com.hedera.mirror.api.proto.ReactorConsensusServiceGrpc;
 import com.hedera.mirror.common.domain.entity.EntityId;
+import com.hedera.mirror.common.domain.entity.EntityType;
+import com.hedera.mirror.common.domain.topic.TopicMessage;
 import com.hedera.mirror.grpc.converter.InstantToLongConverter;
+import com.hedera.mirror.grpc.converter.LongToInstantConverter;
 import com.hedera.mirror.grpc.domain.TopicMessageFilter;
 import com.hedera.mirror.grpc.service.TopicMessageService;
 import com.hedera.mirror.grpc.util.ProtoUtil;
+import com.hederahashgraph.api.proto.java.ConsensusMessageChunkInfo;
 import com.hederahashgraph.api.proto.java.Timestamp;
+import com.hederahashgraph.api.proto.java.TransactionID;
 import java.time.Instant;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -49,7 +55,7 @@ public class ConsensusController extends ReactorConsensusServiceGrpc.ConsensusSe
     public Flux<ConsensusTopicResponse> subscribeTopic(Mono<ConsensusTopicQuery> request) {
         return request.map(this::toFilter)
                 .flatMapMany(topicMessageService::subscribeTopic)
-                .map(topicMessageService::getResponse)
+                .map(this::toResponse)
                 .onErrorMap(ProtoUtil::toStatusRuntimeException);
     }
 
@@ -77,4 +83,55 @@ public class ConsensusController extends ReactorConsensusServiceGrpc.ConsensusSe
 
         return filter.build();
     }
+
+    // These 2 routines are based on the grpc version of TopicMessage.java (before it was dropped to use the version
+    // from hedera-mirror-common, where they were used with a lazy getter to keep the result cached.  Consider
+    // creating a static Map<TopicMessage, ConcensusTopicResponse> in this class to cache the values to save time
+    // regenerating the same ConsensusTopicResponse for multiple subscribers to the same topic.
+    protected ConsensusTopicResponse toResponse(TopicMessage t) {
+        var consensusTopicResponseBuilder = ConsensusTopicResponse.newBuilder()
+                .setConsensusTimestamp(ProtoUtil.toTimestamp(
+                        LongToInstantConverter.INSTANCE.convert(t.getConsensusTimestamp())))
+                .setMessage(ProtoUtil.toByteString(t.getMessage()))
+                .setRunningHash(ProtoUtil.toByteString(t.getRunningHash()))
+                .setRunningHashVersion(t.getRunningHashVersion())
+                .setSequenceNumber(t.getSequenceNumber());
+
+        if (t.getChunkNum() != null) {
+            ConsensusMessageChunkInfo.Builder chunkBuilder = ConsensusMessageChunkInfo.newBuilder()
+                    .setNumber(t.getChunkNum())
+                    .setTotal(t.getChunkTotal());
+
+            TransactionID transactionID = parseTransactionID(t.getInitialTransactionId(),
+                    t.getTopicId().getEntityNum(), t.getSequenceNumber());
+            EntityId payerAccountEntity = t.getPayerAccountId();
+            Instant validStartInstant = LongToInstantConverter.INSTANCE.convert(t.getValidStartTimestamp());
+
+            if (transactionID != null) {
+                chunkBuilder.setInitialTransactionID(transactionID);
+            } else if (payerAccountEntity != null && validStartInstant != null) {
+                chunkBuilder.setInitialTransactionID(TransactionID.newBuilder()
+                        .setAccountID(ProtoUtil.toAccountID(payerAccountEntity))
+                        .setTransactionValidStart(ProtoUtil.toTimestamp(validStartInstant))
+                        .build());
+            }
+
+            consensusTopicResponseBuilder.setChunkInfo(chunkBuilder.build());
+        }
+
+        return consensusTopicResponseBuilder.build();
+    }
+
+    private TransactionID parseTransactionID(byte[] transactionIdBytes, long topicId, long sequenceNumber) {
+        if (transactionIdBytes == null) {
+            return null;
+        }
+        try {
+            return TransactionID.parseFrom(transactionIdBytes);
+        } catch (InvalidProtocolBufferException e) {
+            log.error("Failed to parse TransactionID for topic {} sequence number {}", topicId, sequenceNumber);
+            return null;
+        }
+    }
+
 }
