+++ hedera-mirror-grpc/src/main/java/com/hedera/mirror/grpc/service/TopicMessageService.java
@@ -16,7 +16,6 @@
 
 package com.hedera.mirror.grpc.service;
 
-import com.hedera.mirror.api.proto.ConsensusTopicResponse;
 import com.hedera.mirror.common.domain.topic.TopicMessage;
 import com.hedera.mirror.grpc.domain.TopicMessageFilter;
 import jakarta.validation.Valid;
@@ -25,5 +24,4 @@ import reactor.core.publisher.Flux;
 public interface TopicMessageService {
 
     Flux<TopicMessage> subscribeTopic(@Valid TopicMessageFilter filter);
-    ConsensusTopicResponse getResponse(TopicMessage t);
 }
