+++ hedera-mirror-grpc/src/main/java/com/hedera/mirror/grpc/service/TopicMessageServiceImpl.java
@@ -17,8 +17,6 @@
 package com.hedera.mirror.grpc.service;
 
 import com.google.common.base.Stopwatch;
-import com.google.protobuf.InvalidProtocolBufferException;
-import com.hedera.mirror.api.proto.ConsensusTopicResponse;
 import com.hedera.mirror.common.domain.entity.Entity;
 import com.hedera.mirror.common.domain.entity.EntityId;
 import com.hedera.mirror.common.domain.entity.EntityType;
@@ -30,9 +28,6 @@ import com.hedera.mirror.grpc.exception.EntityNotFoundException;
 import com.hedera.mirror.grpc.listener.TopicListener;
 import com.hedera.mirror.grpc.repository.EntityRepository;
 import com.hedera.mirror.grpc.retriever.TopicMessageRetriever;
-import com.hedera.mirror.grpc.util.ProtoUtil;
-import com.hederahashgraph.api.proto.java.ConsensusMessageChunkInfo;
-import com.hederahashgraph.api.proto.java.TransactionID;
 import io.micrometer.core.instrument.Gauge;
 import io.micrometer.core.instrument.MeterRegistry;
 import jakarta.annotation.PostConstruct;
@@ -95,56 +90,6 @@ public class TopicMessageServiceImpl implements TopicMessageService {
                         .doOnSubscribe(s -> subscriberCount.incrementAndGet())
                         .doFinally(s -> subscriberCount.decrementAndGet())
                         .doFinally(topicContext::finished));
-    }
-
-    // These 2 routines are based on the grpc version of TopicMessage.java (before it was dropped to use the version
-    // from hedera-mirror-common, where they were used with a lazy getter to keep the result cached.  Consider
-    // creating a static Map<TopicMessage, ConcensusTopicResponse> in this class to cache the values to save time
-    // regenerating the same ConsensusTopicResponse for multiple subscribers to the same topic.
-    public ConsensusTopicResponse getResponse(TopicMessage t) {
-        var consensusTopicResponseBuilder = ConsensusTopicResponse.newBuilder()
-                .setConsensusTimestamp(ProtoUtil.toTimestamp(
-                        LongToInstantConverter.INSTANCE.convert(t.getConsensusTimestamp())))
-                .setMessage(ProtoUtil.toByteString(t.getMessage()))
-                .setRunningHash(ProtoUtil.toByteString(t.getRunningHash()))
-                .setRunningHashVersion(t.getRunningHashVersion())
-                .setSequenceNumber(t.getSequenceNumber());
-
-        if (t.getChunkNum() != null) {
-            ConsensusMessageChunkInfo.Builder chunkBuilder = ConsensusMessageChunkInfo.newBuilder()
-                    .setNumber(t.getChunkNum())
-                    .setTotal(t.getChunkTotal());
-
-            TransactionID transactionID = parseTransactionID(t.getInitialTransactionId(),
-                    t.getTopicId().getEntityNum(), t.getSequenceNumber());
-            EntityId payerAccountEntity = t.getPayerAccountId();
-            Instant validStartInstant = LongToInstantConverter.INSTANCE.convert(t.getValidStartTimestamp());
-
-            if (transactionID != null) {
-                chunkBuilder.setInitialTransactionID(transactionID);
-            } else if (payerAccountEntity != null && validStartInstant != null) {
-                chunkBuilder.setInitialTransactionID(TransactionID.newBuilder()
-                        .setAccountID(ProtoUtil.toAccountID(payerAccountEntity))
-                        .setTransactionValidStart(ProtoUtil.toTimestamp(validStartInstant))
-                        .build());
-            }
-
-            consensusTopicResponseBuilder.setChunkInfo(chunkBuilder.build());
-        }
-
-        return consensusTopicResponseBuilder.build();
-    }
-
-    private TransactionID parseTransactionID(byte[] transactionIdBytes, long topicId, long sequenceNumber) {
-        if (transactionIdBytes == null) {
-            return null;
-        }
-        try {
-            return TransactionID.parseFrom(transactionIdBytes);
-        } catch (InvalidProtocolBufferException e) {
-            log.error("Failed to parse TransactionID for topic {} sequence number {}", topicId, sequenceNumber);
-            return null;
-        }
     }
 
     private Mono<?> topicExists(TopicMessageFilter filter) {
